feat: switch to xdm (#9)
* feat: switch to xdm

* things work now

* Update src/index.ts

Co-authored-by: Titus <tituswormer@gmail.com>

Co-authored-by: Titus <tituswormer@gmail.com>

BREAKING CHANGE: We've switched from @mdx-js/js to xdm which is more capable, faster, and requires no runtime! We've also moved to an `xdmOptions` function to customize all the options for compilation rather than simply accepting a `remarkPlugins` option. Check the docs for the latest way to configure things. This is hopefully the last breaking change and we'll enter a pretty stable state from here on out.